+++ src/test/java/org/woehlke/twitterwall/frontend/controller/TweetsControllerTest.java
@@ -1,4 +1,4 @@
-package org.woehlke.twitterwall.controller;
+package org.woehlke.twitterwall.frontend.controller;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -10,7 +10,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.woehlke.twitterwall.Application;
-import org.woehlke.twitterwall.frontend.controller.UserController;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
 
@@ -20,20 +19,19 @@ import static org.assertj.core.api.Java6Assertions.assertThat;
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes={Application.class},webEnvironment=SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
-public class UserControllerTest {
+public class TweetsControllerTest {
 
-    private static final Logger log = LoggerFactory.getLogger(UserControllerTest.class);
+    private static final Logger log = LoggerFactory.getLogger(TweetsControllerTest.class);
     
     @Autowired
     private MockMvc mockMvc;
 
     @Autowired
-    private UserController controller;
-    
+    private TweetsController controller;
+
     @Test
     public void controllerIsPresentTest(){
         log.info("controllerIsPresentTest");
         assertThat(controller).isNotNull();
     }
-    
 }
